+++ src/features/trend/queries/index.ts
@@ -1,0 +1,3 @@
+export { getTrendAlbums, useTrendAlbums } from "./getTrendAlbums";
+export { getTrendArtists, useTrendArtists } from "./getTrendArtists";
+export { getTrendTracks, useTrendTracks } from "./getTrendTracks";
